refactor: move legacy manifest config to uni-builder
