+++ packages/compat/uni-builder/package.json
@@ -26,7 +26,9 @@
   "dependencies": {
     "@rsbuild/core": "workspace:*",
     "@rsbuild/shared": "workspace:*",
-    "@rsbuild/webpack": "workspace:*"
+    "@rsbuild/webpack": "workspace:*",
+    "rspack-manifest-plugin": "5.0.0-alpha0",
+    "webpack-manifest-plugin": "5.0.0"
   },
   "devDependencies": {
     "typescript": "^5.2.2"
