+++ packages/compat/uni-builder/src/webpack/index.ts
@@ -6,19 +6,32 @@ import type {
 } from '@rsbuild/webpack';
 import type { UniBuilderWebpackConfig } from '../types';
 import type { CreateWebpackBuilderOptions } from '../types';
-import { parseCommonConfig } from '../shared';
+import { parseCommonConfig } from '../shared/parseCommonConfig';
 
-export function parseConfig(uniBuilderConfig: UniBuilderWebpackConfig): {
+export async function parseConfig(
+  uniBuilderConfig: UniBuilderWebpackConfig,
+): Promise<{
   rsbuildConfig: RsbuildConfig;
   rsbuildPlugins: RsbuildPlugin[];
-} {
-  return parseCommonConfig<'webpack'>(uniBuilderConfig);
+}> {
+  const { rsbuildConfig, rsbuildPlugins } =
+    parseCommonConfig<'webpack'>(uniBuilderConfig);
+
+  if (uniBuilderConfig.output?.enableAssetManifest) {
+    const { pluginManifest } = await import('./plugins/manifest');
+    rsbuildPlugins.push(pluginManifest());
+  }
+
+  return {
+    rsbuildConfig,
+    rsbuildPlugins,
+  };
 }
 
 export async function createWebpackBuilder(
   options: CreateWebpackBuilderOptions,
 ): Promise<RsbuildInstance<WebpackProvider>> {
-  const { rsbuildConfig, rsbuildPlugins } = parseConfig(options.config);
+  const { rsbuildConfig, rsbuildPlugins } = await parseConfig(options.config);
   const { webpackProvider } = await import('@rsbuild/webpack');
   const rsbuild = await createRsbuild({
     rsbuildConfig,
