+++ packages/compat/uni-builder/src/webpack/plugins/manifest.ts
@@ -1,20 +1,12 @@
-import type { RsbuildPlugin } from '../types';
-import { generateManifest } from '@rsbuild/shared';
+import type { RsbuildPlugin } from '@rsbuild/webpack';
+import { generateManifest } from '../../shared/manifest';
 
 export const pluginManifest = (): RsbuildPlugin => ({
   name: 'plugin-manifest',
 
   setup(api) {
     api.modifyWebpackChain(async (chain, { CHAIN_ID }) => {
-      const config = api.getNormalizedConfig();
-
-      if (!config.output.enableAssetManifest) {
-        return;
-      }
-
-      const { WebpackManifestPlugin } = await import(
-        '../../compiled/webpack-manifest-plugin'
-      );
+      const { WebpackManifestPlugin } = await import('webpack-manifest-plugin');
       const publicPath = chain.output.get('publicPath');
 
       chain.plugin(CHAIN_ID.PLUGIN.MANIFEST).use(WebpackManifestPlugin, [
