+++ packages/core/package.json
@@ -78,7 +78,6 @@
     "open": "^8.4.0",
     "pkg-up": "^3.1.0",
     "postcss": "8.4.31",
-    "rspack-manifest-plugin": "5.0.0-alpha0",
     "semver": "^7.5.4",
     "webpack": "^5.89.0"
   },
