+++ packages/document/docs/en/config/options/output.mdx
@@ -96,12 +96,6 @@ import DisableInlineRuntimeChunk from '@en/shared/config/output/disableInlineRun
 
 <DisableInlineRuntimeChunk />
 
-## output.enableAssetManifest
-
-import EnableAssetManifest from '@en/shared/config/output/enableAssetManifest.md';
-
-<EnableAssetManifest />
-
 ## output.enableAssetFallback
 
 import EnableAssetFallback from '@en/shared/config/output/enableAssetFallback.md';
