+++ packages/document/docs/en/guide/start/features.mdx
@@ -80,7 +80,6 @@ Here are all the main features supported by Rsbuild.
 | Inline static assets   | Small assets are inlined into JS by default                                  | <ul><li>[Inline Static Assets](/guide/optimization/inline-assets.html)</li><li>[output.dataUriLimit](/config/options/output.html#outputdataurilimit)</li></ul> |
 | Clean up static assets | Automatically clean up static assets in the dist directory before each build | <ul><li>[output.cleanDistPath](/config/options/output.html#outputcleandistpath)</li ></ul>                                                                     |
 | Copy static assets     | Optional feature, copy static assets to the dist directory                   | <ul><li>[output.copy](/config/options/output.html#outputcopy)</li></ul>                                                                                        |
-| Generate manifest file | Optional feature, generate manifest file containing static asset information | <ul><li>[output.enableAssetManifest](/config/options/output.html#outputenableassetmanifest)</li></ul>                                                          |
 
 ## Performance
