+++ packages/document/docs/zh/config/options/output.mdx
@@ -96,12 +96,6 @@ import DisableInlineRuntimeChunk from '@zh/shared/config/output/disableInlineRun
 
 <DisableInlineRuntimeChunk />
 
-## output.enableAssetManifest
-
-import EnableAssetManifest from '@zh/shared/config/output/enableAssetManifest.md';
-
-<EnableAssetManifest />
-
 ## output.enableAssetFallback
 
 import EnableAssetFallback from '@zh/shared/config/output/enableAssetFallback.md';
