+++ packages/document/docs/zh/guide/start/features.mdx
@@ -80,7 +80,6 @@
 | 静态资源内联       | 默认将体积较小的图片等资源内联到 JS 中         | <ul><li>[静态资源内联](/guide/optimization/inline-assets.html)</li><li>[output.dataUriLimit](/config/options/output.html#outputdataurilimit)</li></ul> |
 | 清理静态资源       | 每次开始构建前，自动清理 dist 目录下的静态资源 | <ul><li>[output.cleanDistPath](/config/options/output.html#outputcleandistpath)</li></ul>                                                              |
 | 拷贝静态资源       | 可选功能，将静态资源拷贝到 dist 目录下         | <ul><li>[output.copy](/config/options/output.html#outputcopy)</li></ul>                                                                                |
-| 生成 manifest 文件 | 可选功能，生成包含静态资源信息的 manifest 文件 | <ul><li>[output.enableAssetManifest](/config/options/output.html#outputenableassetmanifest)</li></ul>                                                  |
 
 ## 性能相关
